af/requirements.txt
apache-airflow == 2.8.0
apache-airflow-providers-common-sql[openlineage] == 1.8.0
apache-airflow-providers-celery == 3.4.1
apache-airflow-providers-postgres == 5.7.1
apache-airflow-providers-apache-spark == 4.6.0
apache-airflow-providers-apache-hive == 6.4.0
sshtunnel == 0.4.0
paramiko == 3.4.0
apache-airflow-providers-ssh == 3.10.0
